set subscribe handler beforehand
